added jwt token and permissions to create/delete job

+++ apis/v1/route_jobs.py
@@ -1,12 +1,15 @@
 from typing import List
+
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.orm import Session
 
+from apis.v1.route_login import get_current_user_from_token
+from db.models.users import User
 from db.repository.jobs import (
     create_new_job,
     delete_job_by_id,
-    retrieve_job,
     list_jobs,
+    retrieve_job,
     update_job_by_id,
 )
 from db.session import get_db
@@ -16,9 +19,13 @@ router = APIRouter()
 
 
 @router.post('/create-job', response_model=ShowJob)
-def create_job(job: JobCreate, db: Session = Depends(get_db)):
-    owner_id = 1
-    job = create_new_job(job, db, owner_id)
+def create_job(
+    job: JobCreate,
+    db: Session = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
+):
+    owner_id = current_user.id
+    job = create_new_job(job=job, db=db, owner_id=owner_id)
     return job
 
 
@@ -54,11 +61,21 @@ def update_job(id: int, job: JobCreate, db: Session = Depends(get_db)):
 
 
 @router.delete('/delete/{id}')
-def delete_job(id: int, db: Session = Depends(get_db)):
-    message = delete_job_by_id(id=id, db=db)
-    if not message:
+def delete_job(
+    id: int,
+    db: Session = Depends(get_db),
+    current_user: User = Depends(get_current_user_from_token),
+):
+    job = retrieve_job(id=id, db=db)
+    if not job:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f'Job with id {id} doest not exist',
         )
-    return {'detail': f'Successfully deleted the job with id {id}'}
+    if job.owner_id == current_user.id or current_user:
+        delete_job_by_id(id=id, db=db)
+        return {'detail': 'Job successfully deleted'}
+
+    raise HTTPException(
+        status=status.HTTP_401_UNAUTHORIZED, detail='You are not permitted'
+    )
